ghaith

+++ database/migrations/2025_12_09_054148_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('product_id')->constrained('products');
-            $table->Integer('rating'); 
+            $table->float('rating'); 
             $table->text('comment')->nullable();
             $table->timestamps();
         });
